INT-3384: fix test failure
